consolidating gulpfile, linting
