+++ demos/index.js
@@ -1,5 +1,7 @@
 import 'angular';
-import '../src/dataTable.js';
+
+import '../src/dataTable';
+
 import '../src/dataTable.less';
 import '../src/themes/icons.less';
 import '../src/themes/material.less';
